src/pinky_pro/pinky_llm/requirements.txt
langchain
langchain-openai
langchain-core
langchain-community
pyyaml
python-dotenv
